+++ Libplanet.Explorer/GraphTypes/TxResult.cs
@@ -1,8 +1,4 @@
-using System.Collections.Immutable;
-using Bencodex.Types;
-using Libplanet.Crypto;
-using Libplanet.Types.Assets;
-using FAV = Libplanet.Types.Assets.FungibleAssetValue;
+using System.Collections.Generic;
 
 namespace Libplanet.Explorer.GraphTypes
 {
@@ -12,18 +8,12 @@ namespace Libplanet.Explorer.GraphTypes
             TxStatus status,
             long? blockIndex,
             string? blockHash,
-            string? exceptionName,
-            IImmutableDictionary<Address, IValue>? updatedStates,
-            IImmutableDictionary<Address, IImmutableDictionary<Currency, FAV>>?
-                updatedFungibleAssets
-        )
+            List<string?>? exceptionNames)
         {
             TxStatus = status;
             BlockIndex = blockIndex;
             BlockHash = blockHash;
-            ExceptionName = exceptionName;
-            UpdatedStates = updatedStates;
-            UpdatedFungibleAssets = updatedFungibleAssets;
+            ExceptionNames = exceptionNames;
         }
 
         public TxStatus TxStatus { get; private set; }
@@ -32,11 +22,6 @@ namespace Libplanet.Explorer.GraphTypes
 
         public string? BlockHash { get; private set; }
 
-        public string? ExceptionName { get; private set; }
-
-        public IImmutableDictionary<Address, IValue>? UpdatedStates { get; }
-
-        public IImmutableDictionary<Address, IImmutableDictionary<Currency, FungibleAssetValue>>?
-            UpdatedFungibleAssets { get; }
+        public List<string?>? ExceptionNames { get; private set; }
     }
 }